Move ConsensusNotifier to new module
This will need to be used by other components beyond the chain controller.
