+++ validator/src/consensus/mod.rs
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 Intel Corporation
+ * Copyright 2018 Cargill Incorporated
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,39 +15,5 @@
  * ------------------------------------------------------------------------------
  */
 
-extern crate cbor;
-extern crate cpython;
-extern crate hex;
-extern crate libc;
-extern crate lmdb_zero;
-extern crate protobuf;
-extern crate python3_sys as py_ffi;
-#[macro_use]
-extern crate lazy_static;
-#[macro_use]
-extern crate log;
-extern crate openssl;
-#[cfg(test)]
-extern crate rand;
-extern crate uluru;
-
-// exported modules
-pub mod consensus;
-pub mod database;
-pub mod execution;
-pub mod gossip;
-pub mod hashlib;
-pub mod journal;
-mod metrics;
-pub mod proto;
-pub mod pylogger;
-pub mod scheduler;
-pub mod state;
-
-pub mod batch;
-mod batch_ffi;
-pub mod block;
-mod block_ffi;
-pub mod transaction;
-
-pub mod ffi;
+pub mod notifier;
+pub mod notifier_ffi;
