+++ validator/src/consensus/notifier.rs
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 Intel Corporation
+ * Copyright 2018 Cargill Incorporated
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,39 +15,11 @@
  * ------------------------------------------------------------------------------
  */
 
-extern crate cbor;
-extern crate cpython;
-extern crate hex;
-extern crate libc;
-extern crate lmdb_zero;
-extern crate protobuf;
-extern crate python3_sys as py_ffi;
-#[macro_use]
-extern crate lazy_static;
-#[macro_use]
-extern crate log;
-extern crate openssl;
-#[cfg(test)]
-extern crate rand;
-extern crate uluru;
+use block::Block;
 
-// exported modules
-pub mod consensus;
-pub mod database;
-pub mod execution;
-pub mod gossip;
-pub mod hashlib;
-pub mod journal;
-mod metrics;
-pub mod proto;
-pub mod pylogger;
-pub mod scheduler;
-pub mod state;
-
-pub mod batch;
-mod batch_ffi;
-pub mod block;
-mod block_ffi;
-pub mod transaction;
-
-pub mod ffi;
+pub trait ConsensusNotifier: Send + Sync {
+    fn notify_block_new(&self, block: &Block);
+    fn notify_block_valid(&self, block_id: &str);
+    fn notify_block_invalid(&self, block_id: &str);
+    fn notify_block_commit(&self, block_id: &str);
+}
